Validate Transactions
For security purposes, We need to validate transactions before we can record them in the database. For guaranteed accuracy we use the 'GetTransaction' endpoints to confirm that a transaction did go through. Using $_GET variables exposes the data and a user can simply edit the variables in the url and cause us harm.